// web/src/api/generated/models/exam-template-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 试卷模板响应
 */
export interface ExamTemplateResponse {
    /**
     * 模板ID
     */
    'id'?: number;
    /**
     * 模板名称
     */
    'name'?: string;
    /**
     * 模板描述
     */
    'description'?: string;
    /**
     * 模板文件URL
     */
    'templateUrl'?: string;
    /**
     * 预览封面URL
     */
    'coverUrl'?: string;
    /**
     * 是否系统内置
     */
    'isSystem'?: boolean;
    /**
     * 是否启用
     */
    'isEnabled'?: boolean;
    /**
     * 创建者ID
     */
    'creatorId'?: number;
    /**
     * 创建时间
     */
    'createTime'?: string;
    /**
     * 更新时间
     */
    'updateTime'?: string;
}